+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
 
 
+## Since 0.2.0 (26 Aug 2017):
+- Improve display selected word
+- Added execute command `removeSecondaryCursors` to remove multiple selections
+
 ## Since 0.1.0 (25 Aug 2017):
 - Initial release
 - Added `goToWord.find` command
-- Added execute command `addSelectionToNextFindMatch` after select word
+- Added execute command `editor.action.addSelectionToNextFindMatch` after select word